% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_monthly_excess.r
\name{estimate_monthly_excess}
\alias{estimate_monthly_excess}
\title{Estimate excess mortality using monthly data}
\usage{
estimate_monthly_excess(
  outcome_variable,
  data,
  forecast_window = 14,
  forecast_start = as.Date("2020-03-01"),
  forecast_periods = NULL,
  data_start = c(2016, 1)
)
}
\arguments{
\item{outcome_variable}{The name of the outcome variable, in quotes.}

\item{data}{A data frame of the aggregated data. It should have a \code{Date} variable named \code{month} and at least one variable that represents the number of deaths per month within some population (see \code{outcome_variable} above).}

\item{forecast_window}{The number of months in the forecast window.}

\item{forecast_start}{The start date of the forecast window, in Date format.}

\item{forecast_periods}{Time periods (optional) within the forecast period. This should be specified as a numeric vector that has a length matching the forecast period. The values of the period should identify the periods. For example: \code{c(rep(1,5),rep(2,5),rep(3,6))}. If validly specified, the function will return results for each time period, in addition to results for the entire forecast window.}

\item{data_start}{The start date of the data, specified as a two-element numeric vector. See \code{ts()}.}
}
\value{
A list with five elements:
\itemize{
\item \code{results_by_month}: The monthly results.
\item \code{results}: The overall results (that is, over the forecast window).
\item \code{simulations}: Simulated sums, used to obtain the prediction intervals in \code{results} (above).
\item \code{plot}: A visualization of the results.
\item \code{plot_data}: The data underlying the plot.
}
}
\description{
Estimate excess mortality using monthly data
}
\examples{
rr<-estimate_monthly_excess('Inland Empire',monthly_data)
rr$results
}
