% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_weekly_excess.r
\name{estimate_weekly_excess}
\alias{estimate_weekly_excess}
\title{Estimate excess mortality using weekly data}
\usage{
estimate_weekly_excess(
  outcome_variable,
  data,
  forecast_window = 143,
  forecast_start = as.Date("2020-03-07"),
  forecast_periods = NULL,
  data_start = c(2016, 15/7)
)
}
\arguments{
\item{outcome_variable}{The name of the outcome variable, in quotes.}

\item{data}{A data frame of the aggregated data. It should have a \code{Date} variable named \code{week} and at least one variable that represents the number of deaths per week within some population (see \code{outcome_variable} above). \code{week} should represent the date of the last day of each week (for example, if you are using US convention, the last day of each week should be a Saturday).}

\item{forecast_window}{The number of weeks in the forecast window.}

\item{forecast_start}{The start date of the forecast window, in Date format.}

\item{forecast_periods}{Time periods (optional) within the forecast period. This should be specified as a numeric vector that has a length matching the forecast period. The values of the period should identify the periods. For example: \code{c(rep(1,12),rep(2,8),rep(3,15))}. If validly specified, the function will return results for each time period, in addition to results for the entire forecast window.}

\item{data_start}{The start date of the data, specified as a two-element numeric vector. See \code{ts()}. For example, \code{c(2016,15/7)} is the week ending 2016-01-09. That's because 2016-01-01 would be \code{c(2016,7/7)} and 2016-01-08 would be \code{c(2016,14/7)}..}
}
\value{
A list with five elements:
\itemize{
\item \code{results_by_week}: The weekly results.
\item \code{results}: The overall results (that is, over the forecast window).
\item \code{simulations}: Simulated sums, used to obtain the prediction intervals in \code{results} (above).
\item \code{plot}: A visualization of the results.
\item \code{plot_data}: The data underlying the plot.
}
}
\description{
Estimate excess mortality using weekly data
}
\examples{
rr<-estimate_weekly_excess('Inland Empire',weekly_data)
rr$results
}
